pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "speclink"
version = "0.1.1"
description = "Keep documentation and code in sync using LLMs"
authors = [
    { name = "lacatu5" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
keywords = ["documentation", "sync", "llm", "ai", "development-tools"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Documentation",
]
dependencies = [
    "beaupy",
    "instructor",
    "litellm",
    "markdown-it-py",
    "pathspec",
    "pydantic>=2",
    "pydantic-settings",
    "PyYAML",
    "structlog",
    "tiktoken",
    "tree-sitter",
    "tree-sitter-python",
    "tree-sitter-typescript",
    "typer",
]

[project.urls]
Homepage = "https://github.com/lacatu5/speclink"
Issues = "https://github.com/lacatu5/speclink/issues"

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]

[project.scripts]
speclink = "speclink.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.12"
strict = false

[[tool.mypy.overrides]]
module = [
    "tree_sitter.*",
    "tree_sitter_language_pack.*",
    "litellm.*",
    "instructor.*",
    "beaupy.*",
    "pathspec.*",
    "structlog.*",
    "tiktoken.*",
]
ignore_missing_imports = true
